requirements.txt
Flask==2.3.3  
Flask-Login==0.6.3  
Flask-SQLAlchemy==3.1.1  
argon2-cffi==23.1.0  